+++ docs/storybook/stories/components/Modal.stories.tsx
@@ -1,8 +1,8 @@
-import * as React from 'react';
-import { Box, Button, Flex, Text } from '@ttoss/ui';
-import { Meta, StoryFn } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react-webpack5';
 import { Modal } from '@ttoss/components/Modal';
-import { action } from '@storybook/addon-actions';
+import { Box, Button, Flex, Text } from '@ttoss/ui';
+import * as React from 'react';
+import { action } from 'storybook/actions';
 
 export default {
   title: 'Components/Modal',