+++ docs/storybook/stories/forms/Form.stories.tsx
@@ -1,6 +1,5 @@
 import alertIcon from '@iconify-icons/mdi-light/alert';
-import { action } from '@storybook/addon-actions';
-import { Meta, StoryFn } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react-webpack5';
 import {
   Form,
   FormFieldCheckbox,
@@ -18,6 +17,7 @@ import {
 import { I18nProvider } from '@ttoss/react-i18n';
 import { Box, Button, Flex } from '@ttoss/ui';
 import * as React from 'react';
+import { action } from 'storybook/actions';
 
 const loadLocaleData = async (locale: string) => {
   switch (locale) {