+++ docs/storybook/stories/react-auth-core/AuthSignIn.stories.tsx
@@ -1,11 +1,11 @@
-import { Meta, StoryFn } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react-webpack5';
 import { Box } from '@ttoss/ui';
 
-import { LogoProvider } from '../../../../packages/react-auth/src/AuthCard';
+import { LogoProvider } from '../../../../packages/react-auth-core/src/AuthCard';
 import {
   AuthSignIn,
   AuthSignInProps,
-} from '../../../../packages/react-auth/src/AuthSignIn';
+} from '../../../../packages/react-auth-core/src/AuthSignIn';
 
 export default {
   title: 'React Auth/AuthSignIn',