+++ docs/storybook/stories/ui/Card.stories.tsx
@@ -1,4 +1,4 @@
-import { Meta, StoryFn } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react-webpack5';
 import { Card, CardProps, Flex, Text } from '@ttoss/ui';
 
 export default {