+++ docs/storybook/stories/ui/Label.stories.tsx
@@ -1,6 +1,6 @@
-import { action } from '@storybook/addon-actions';
-import { Meta, StoryObj } from '@storybook/react';
+import { Meta, StoryObj } from '@storybook/react-webpack5';
 import { Label, LabelProps } from '@ttoss/ui';
+import { action } from 'storybook/actions';
 
 type Story = StoryObj<typeof Label>;