+++ docs/storybook/stories/ui/Switch.stories.tsx
@@ -1,4 +1,4 @@
-import { Meta, StoryFn } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react-webpack5';
 import { type CheckboxProps, Switch } from '@ttoss/ui';
 
 export default {