+++ packages/react-auth-core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-core@0.2.1...@ttoss/react-auth-core@0.2.2) (2025-09-19)
+
+**Note:** Version bump only for package @ttoss/react-auth-core
+
 ## [0.2.1](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-core@0.2.0...@ttoss/react-auth-core@0.2.1) (2025-08-16)
 
 ### Bug Fixes